+++ src/api/java/mekanism/api/chemical/infuse/IMekanismInfusionHandler.java
@@ -78,7 +78,7 @@ public interface IMekanismInfusionHandler extends ISidedInfusionHandler {
     }
 
     @Override
-    default int getInfusionTankCapacity(int tank, @Nullable Direction side) {
+    default long getInfusionTankCapacity(int tank, @Nullable Direction side) {
         IChemicalTank<InfuseType, InfusionStack> infusionTank = getInfusionTank(tank, side);
         return infusionTank == null ? 0 : infusionTank.getCapacity();
     }
@@ -96,7 +96,7 @@ public interface IMekanismInfusionHandler extends ISidedInfusionHandler {
     }
 
     @Override
-    default InfusionStack extractInfusion(int tank, int amount, @Nullable Direction side, Action action) {
+    default InfusionStack extractInfusion(int tank, long amount, @Nullable Direction side, Action action) {
         IChemicalTank<InfuseType, InfusionStack> infusionTank = getInfusionTank(tank, side);
         return infusionTank == null ? InfusionStack.EMPTY : infusionTank.extract(amount, action, side == null ? AutomationType.INTERNAL : AutomationType.EXTERNAL);
     }
