+++ src/main/java/mekanism/common/content/boiler/BoilerCache.java
@@ -27,7 +27,7 @@ public class BoilerCache extends MultiblockCache<SynchronizedBoilerData> impleme
 
     //Note: We don't care about any restrictions here as it is just for making it be persistent
     private final List<IExtendedFluidTank> fluidTanks = Collections.singletonList(BasicFluidTank.create(Integer.MAX_VALUE, this));
-    private final List<IChemicalTank<Gas, GasStack>> gasTanks = Collections.singletonList(BasicGasTank.create(Integer.MAX_VALUE, this));
+    private final List<IChemicalTank<Gas, GasStack>> gasTanks = Collections.singletonList(BasicGasTank.create(Long.MAX_VALUE, this));
     public final List<IHeatCapacitor> heatCapacitors = Collections.singletonList(BasicHeatCapacitor.create(HeatAPI.DEFAULT_HEAT_CAPACITY, this));
 
     @Override
